split candidate validation metrics
